HNSW-E2E: 1. Removed obsolete module.
