+++ iris-mpc-upgrade-hawk/tests/workflows/genesis/mod.rs
@@ -1,4 +1,3 @@
-pub(crate) mod factory;
 pub(crate) mod inputs;
 pub(crate) mod net;
 pub(crate) mod params;
